+++ apps/backyard/src/model/domain/championship.ts
@@ -3,7 +3,7 @@ import { BaseModel } from '@/firebase/db/base-model';
 export interface Championship extends BaseModel {
   title: string;
   nr: number;
-  rulesetId: string;
+  rulesId: string;
   published: boolean;
   completed: boolean;
 }